# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bitcoin Mining Farm Calculator API",
    "version": "1.0.0",
    "description": "Free API for calculating Bitcoin mining farm profitability, CAPEX, OPEX, and multi-year revenue forecasts. Designed for developers, AI agents, and mining operators. Rate limited to 60 requests/minute for external callers.",
    "contact": {
      "name": "Bitcoin Mining Farm Calculator",
      "url": "https://bitcoinminingfarmcalculator.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://bitcoinminingfarmcalculator.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
...
      "CalculateResponse": {
        "type": "object",
        "properties": {
          "metrics": {
            "type": "object",
            "properties": {
              "totalHashRateThs": {
                "type": "number",
                "description": "* Total wall power draw of all miners (kW)"
              },
              "totalPowerKw": {
                "type": "number",
                "description": "* Estimated monthly energy consumption (kWh)"
              },
              "monthlyKwh": {
                "type": "number",
                "description": "* Total heat output (BTU/h)"
              },
              "heatOutputBtuPerHour": {
                "type": "number",
                "description": "* Total electrical current draw (A)"
              },
              "electricalCurrentAmps": {
                "type": "number",
                "description": "* Required transformer capacity (kVA)"
              },
              "transformerKva": {
                "type": "number",
                "description": "* Transformer purchase cost (USD)"
              },
              "transformerCost": {
                "type": "number",
                "description": "* Total copper cable cost (USD)"
              },
              "cableCost": {
                "type": "number",
                "description": "* Rack hardware cost (USD, 0 if using containers)"
              },
              "rackCost": {
                "type": "number",
                "description": "* Container hardware cost (USD, 0 if using racks)"
              },
              "containerCost": {
                "type": "number",
                "description": "* Cooling infrastructure cost (USD)"
              },
              "coolingCost": {
                "type": "number",
                "description": "* Solar panel system CAPEX (USD)"
              },
              "solarCapex": {
                "type": "number",
                "description": "* Total deployment labor cost (USD)"
              },
              "laborCost": {
                "type": "number",
                "description": "* Total cables and breakers cost (USD)"
              },
              "cablesAndBreakers": {
                "type": "number",
                "description": "* Dry cooler equipment CAPEX (USD)"
              },
              "dryCoolerCapex": {
                "type": "number",
                "description": "* Air fan equipment CAPEX (USD)"
              },
              "airFanCapex": {
                "type": "number",
                "description": "* Total power consumed by air fans (kW)"
              },
              "airFanPowerKw": {
                "type": "number",
                "description": "* Total miner hardware cost (USD)"
              },
              "minerCost": {
                "type": "number",
                "description": "* Total import tax on all hardware (USD)"
              },
              "importTaxCapex": {
                "type": "number",
                "description": "* Grand total capital expenditure (USD)"
              },
              "totalCapex": {
                "type": "number",
                "description": "* Total monthly operating expenditure (USD)"
              },
              "monthlyOpex": {
                "type": "number",
                "description": "* Monthly solar maintenance cost (USD)"
              },
              "monthlySolarMaintenance": {
                "type": "number",
                "description": "* Monthly maintenance labor cost (USD)"
              },
              "maintenanceLaborOpex": {
                "type": "number"
              }
            },
            "required": [
              "totalHashRateThs",
              "totalPowerKw",
              "monthlyKwh",
              "heatOutputBtuPerHour",
              "electricalCurrentAmps",
              "transformerKva",
              "transformerCost",
              "cableCost",
              "rackCost",
              "containerCost",
              "coolingCost",
              "solarCapex",
              "laborCost",
              "cablesAndBreakers",
              "dryCoolerCapex",
              "airFanCapex",
              "airFanPowerKw",
              "minerCost",
              "importTaxCapex",
              "totalCapex",
              "monthlyOpex",
              "monthlySolarMaintenance",
              "maintenanceLaborOpex"
            ],
            "description": "* Required ventilation airflow"
          },
          "ventilation": {
            "type": "object",
            "properties": {
              "m3h": {
                "type": "number",
                "description": "* Airflow in cubic feet per minute"
              },
              "cfm": {
                "type": "number"
              }
            },
            "required": [
              "m3h",
              "cfm"
            ],
            "description": "* Combined hash rate of all miners (TH/s)"
          },
          "totalHashRateThs": {
            "type": "number",
            "description": "* Combined wall power of all miners (kW)"
          },
          "totalPowerKw": {
            "type": "number",
            "description": "* Total power consumed by air cooling fans (kW)"
          },
          "airFanPowerKw": {
            "type": "number"
          }
        },
        "required": [
          "metrics",
          "ventilation",
          "totalHashRateThs",
          "totalPowerKw",
          "airFanPowerKw"
        ]
      },
...
      "ForecastBody": {
        "type": "object",
        "properties": {
          "config": {
            "description": "* Forecast simulation parameters",
            "nullable": false
          },
          "params": {
            "type": "object",
            "properties": {
              "months": {
                "type": "number",
                "enum": [
                  12,
                  24,
                  36,
                  48,
                  72
                ],
                "description": "* Revenue strategy: sell all BTC immediately, hold all BTC, or sell only enough to cover OPEX",
                "nullable": false
              },
              "revenueMode": {
                "type": "string",
                "enum": [
                  "sell_all",
                  "hold_all",
                  "sell_opex"
                ],
                "description": "* BTC price projection model — \"fixed\" holds price constant, \"stock_to_flow\" uses S2F curve, \"stock_to_flow_pessimistic\" applies a haircut, \"custom\" uses startingBtcPrice→finalBtcPrice linear interpolation",
                "nullable": false
              },
              "btcPriceModel": {
                "type": "string",
                "enum": [
                  "fixed",
                  "stock_to_flow",
                  "stock_to_flow_pessimistic",
                  "custom"
                ],
                "description": "* Pessimistic discount applied to the S2F model (negative percent, e.g. -30 means 30% below S2F)",
                "nullable": false
              },
              "pessimisticAdjustPercent": {
                "type": "number",
                "description": "* Annual network hashrate growth rate (percent, e.g. 25 for 25% YoY)",
                "nullable": false
              },
              "networkHashrateGrowthPercent": {
                "type": "number",
                "description": "* Annual ASIC performance degradation (percent, e.g. 5 for 5% YoY hashrate loss)",
                "nullable": false
              },
              "asicDegradationPercent": {
                "type": "number",
                "description": "* Annual discount rate for NPV / IRR calculations (percent, default 10)",
                "nullable": false
              },
              "discountRatePercent": {
                "type": "number",
                "description": "* Starting BTC price in USD (current market price)",
                "nullable": false
              },
              "startingBtcPrice": {
                "type": "number",
                "description": "* Final BTC price override in USD — null lets the model calculate it automatically",
                "nullable": false
              },
              "finalBtcPrice": {
                "type": "number",
                "nullable": false
              }
            },
            "required": [
              "months",
              "revenueMode",
              "btcPriceModel",
              "pessimisticAdjustPercent",
              "networkHashrateGrowthPercent",
              "asicDegradationPercent",
              "discountRatePercent",
              "startingBtcPrice",
              "finalBtcPrice"
            ],
            "nullable": false
          }
        },
        "required": [
          "config",
          "params"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request — invalid or missing request body",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                }
              }
            }
          }
        }
      },
      "429": {
        "description": "Too Many Requests — rate limit exceeded (60 req/min for external callers)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Too many requests"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/calculate": {
      "post": {
        "operationId": "post-calculate",
        "summary": "Calculate farm metrics",
        "description": "Calculates comprehensive CAPEX, OPEX, and operational metrics for a Bitcoin mining farm configuration. Accepts a complete FarmConfig and returns all derived values including total hash rate, power draw, heat output, transformer sizing, infrastructure costs, and monthly operating expenses. This is the primary calculation endpoint — send your full farm configuration and receive all metrics in one call.",
        "tags": [
          "Calculate"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculateResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/forecast": {
      "post": {
        "operationId": "post-forecast",
        "summary": "Generate multi-year revenue forecast",
        "description": "Generates a month-by-month Bitcoin mining revenue forecast with BTC price modeling (Stock-to-Flow), network difficulty growth, ASIC degradation, energy inflation, and halving events. Returns NPV, IRR, break-even BTC price, payback period, and cumulative profit/loss for each period. Supports three revenue strategies: sell all BTC, hold all BTC, or sell only enough to cover OPEX.",
        "tags": [
          "Forecast"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}